Move test directories to more decentralized folders
