+++ src/modules/http/__tests__/axios.service.test.ts
@@ -1,4 +1,4 @@
-import { AxiosService } from '../http/axios.service';
+import { AxiosService } from '../axios.service';
 
 describe('Axios Service', () => {
     describe('Perform get request', () => {
